Uploaded file and folder finder
The first version of the file and folder finder, which is portable i.e. working on three different OSes (Windows, Linux, macOS). Its search space is all files(or folders) present on the pc.

**Features**: 
(1) Case-Insensitive Search of file or folder name.
(2) If there is a unique file with a given name, then it will find and open it. 
     Otherwise, it will print all the paths of found files(or folders) with the same name.